+++ src/main/kotlin/app/civa/vaccination/adapter/in/VaccinationCardPayloadIn.kt
@@ -3,4 +3,4 @@ package app.civa.vaccination.adapter.`in`
 import app.civa.vaccination.domain.Species
 import java.util.*
 
-data class VaccinationCardPayloadIn(val petId: UUID, val species: Species) {}
+data class VaccinationCardPayloadIn(val petId: UUID, val species: Species)
