+++ src/main/kotlin/app/civa/vaccination/adapter/in/VaccinationCardPayloadOut.kt
@@ -1,8 +1,6 @@
 package app.civa.vaccination.adapter.`in`
 
-import app.civa.vaccination.domain.Applications
-import app.civa.vaccination.domain.Species
-import app.civa.vaccination.domain.VaccinationCard
+import app.civa.vaccination.domain.*
 import java.util.*
 
 data class VaccinationCardPayloadOut(
@@ -11,19 +9,36 @@ data class VaccinationCardPayloadOut(
     val species: Species,
     val applications: Applications
 ) {
-    constructor(builder: VaccinationCardPayloadOutBuilder) : this(
-        builder.id,
-        builder.petID,
-        builder.species,
-        builder.applications
-    )
-
     companion object {
+
         infix fun from(vaccinationCard: VaccinationCard): VaccinationCardPayloadOut {
-            val builder = VaccinationCardPayloadOutBuilder()
+            val builder = object : VaccinationCardVisitor, Builder<VaccinationCardPayloadOut> {
+                lateinit var id: UUID
+                lateinit var petID: UUID
+                lateinit var species: Species
+                lateinit var applications: Applications
+
+                override fun seeId(id: UUID) {
+                    this.id = id
+                }
+
+                override fun seePetId(petID: UUID) {
+                    this.petID = petID
+                }
+
+                override fun seeSpecies(species: Species) {
+                    this.species = species
+                }
+
+                override fun seeApplications(applications: Applications) {
+                    this.applications = applications
+                }
+
+                override fun build() = VaccinationCardPayloadOut(id, petID, species, applications)
+            }
+
             vaccinationCard accepts builder
             return builder.build()
         }
     }
 }
-
